Update snapshots, refactor to util
