+++ src/v3/src/util/languageUtils.ts
@@ -10,6 +10,8 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
+import { OdysseyI18nResourceKeys, odysseyI18nResourceKeysList } from '@okta/odyssey-react-mui';
+
 import Bundles from '../../../util/Bundles';
 import { WidgetProps } from '../types';
 import { getLanguageCode, getSupportedLanguages } from './settingsUtils';
@@ -24,3 +26,15 @@ export const loadLanguage = async (widgetProps: WidgetProps): Promise<void> => {
     rewrite: rewrite ?? ((val) => val),
   }, supportedLanguages);
 };
+
+export const getOdysseyTranslationOverrides = (): Partial<OdysseyI18nResourceKeys> => (
+  odysseyI18nResourceKeysList
+    .reduce((overrides: Partial<OdysseyI18nResourceKeys>,
+      key: typeof odysseyI18nResourceKeysList[number]) => {
+      const updatedOverrides = { ...overrides };
+      if (Bundles.login && Object.prototype.hasOwnProperty.call(Bundles.login, key)) {
+        updatedOverrides[key] = (Bundles.login as Partial<OdysseyI18nResourceKeys>)[key];
+      }
+      return updatedOverrides;
+    }, {})
+);
